move the remainder of general testing utils to internal/
I missed mocks3 and tc in #4697.  And tchttputil is utilities, so a good
fit for internal/
