+++ internal/mocktc/queueprovider.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/taskcluster/taskcluster/v43/clients/client-go/tcqueue"
-	"github.com/taskcluster/taskcluster/v43/workers/generic-worker/tc"
+	"github.com/taskcluster/taskcluster/v43/internal/mocktc/tc"
 )
 
 type QueueProvider struct {
